perf(rust): move pid, state, profile, rate_monitor to f32
Continues the f32 migration into the control core, with NO integer-time
rework: time stays f64 only as anchors. The large absolute `now` /
start_time / prev_time values remain f64, and only the small dt / elapsed
*deltas* derived from them are narrowed to f32 (the subtraction is done in
f64 first, so there is no cancellation regardless of uptime).

- pid: gains, error, P/I/D terms and the integral are f32; `now`/`prev_time`
  stay f64 and dt is computed in f64 then narrowed. The integral keeps its
  anti-windup hard clamp, which is what makes f32 accumulation safe.
- state: temps, targets and elapsed-relative time (elapsed_offset,
  step_start_time, stall/record timers) are f32. KilnError and
  ControllerConfig stay f64 (wire/config format), cast at the few use sites;
  the max-temp safety compare is still performed in f64.
- profile / rate_monitor: fully f32 (bounded, no accumulation). TempHistory
  now stores (time, temp) as f32 -> 8 B/sample instead of 16.

The f32<->f64 boundary casts the leaf round introduced are gone now that pid
and state speak f32. protocol/json/csv stay f64, with widening confined to
build_status and the resume/tuning command boundaries (the tuner is still
f64). Golden replay tolerances were relaxed for the f32 paths; the discrete
outputs (state, step index, integral-freeze flag) still match the f64
reference exactly.

Measured (stable build, real binary) vs the leaf-round baseline:
  RAM (.data+.bss): 206,900 -> 203,396 B   (-3,504; control_task pool -1,752)
  Flash:            750,896 -> 747,912 B   (-2,984; the leaf round's
                                            boundary-cast tax is recovered)

+++ rust/kiln-app/src/profile_json.rs
@@ -216,10 +216,12 @@ impl<'a> Reader<'a> {
     fn parse_step(&mut self) -> Result<Step, ProfileJsonError> {
         self.expect(b'{')?;
         let mut kind: Option<StepKind> = None;
-        let mut target_temp: Option<f64> = None;
-        let mut desired_rate: Option<f64> = None;
-        let mut min_rate: Option<f64> = None;
-        let mut duration: Option<f64> = None;
+        // `Step` stores f32; parse as f64 then narrow (kiln temps/rates/durations
+        // are far inside f32's exact range).
+        let mut target_temp: Option<f32> = None;
+        let mut desired_rate: Option<f32> = None;
+        let mut min_rate: Option<f32> = None;
+        let mut duration: Option<f32> = None;
 
         self.skip_ws();
         if !self.try_consume(b'}') {
@@ -238,10 +240,10 @@ impl<'a> Reader<'a> {
                             _ => return Err(ProfileJsonError::BadStepType),
                         });
                     }
-                    "target_temp" => target_temp = Some(self.parse_number()?),
-                    "desired_rate" => desired_rate = Some(self.parse_number()?),
-                    "min_rate" => min_rate = Some(self.parse_number()?),
-                    "duration" => duration = Some(self.parse_number()?),
+                    "target_temp" => target_temp = Some(self.parse_number()? as f32),
+                    "desired_rate" => desired_rate = Some(self.parse_number()? as f32),
+                    "min_rate" => min_rate = Some(self.parse_number()? as f32),
+                    "duration" => duration = Some(self.parse_number()? as f32),
                     _ => self.skip_value(0)?,
                 }
                 self.skip_ws();
